requirements.txt
altgraph==0.17.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
distlib==0.3.4
docopt==0.6.2
docutils==0.18.1
filelock==3.4.2
future==0.18.2
idna==3.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.2
ldap3==2.9.1
Naked==0.1.31
numpy==1.22.0
pandas==1.3.5
pefile==2021.9.3
Pillow==9.0.0
pipreqs==0.4.11
platformdirs==2.4.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.12.0
pydirectory==0.1.15
Pygments==2.11.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.0
rsa==4.8
shellescape==3.8.1
six==1.16.0
urllib3==1.26.7
virtualenv==20.13.0
yarg==0.1.9
